@@ -16,12 +16,19 @@
     "e2e:test": "detox test --configuration ios.sim.debug"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.9",
+    "@react-navigation/bottom-tabs": "^5.2.6",
+    "@react-navigation/drawer": "^5.4.1",
+    "@react-navigation/native": "^5.1.5",
+    "@react-navigation/stack": "^5.2.10",
+    "global": "^4.4.0",
     "i18n-js": "^3.2.1",
+    "jetifier": "^1.6.5",
     "lodash": "^4.17.13",
     "moment": "^2.24.0",
-    "react": "16.8.1",
-    "react-native": "0.61.3",
-    "react-native-calendars": "^1.212.0",
+    "react": "16.9.0",
+    "react-native": "0.61.5",
+    "react-native-calendars": "1.265.0",
     "react-native-gesture-handler": "^1.5.0",
     "react-native-gifted-chat": "^0.11.3",
     "react-native-lightbox": "^0.8.0",
@@ -29,13 +36,13 @@
     "react-native-localize": "^1.3.1",
     "react-native-modal": "^9.0.0",
     "react-native-modal-datetime-picker": "^6.0.0",
-    "react-native-modal-dropdown": "^0.6.2",
+    "react-native-modal-dropdown": "0.7.0",
     "react-native-paper": "^3.1.1",
+    "react-native-reanimated": "^1.8.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.5.0",
     "react-native-ui-lib": "^3.39.0",
     "react-native-vector-icons": "^6.6.0",
-    "react-navigation": "^3.3.2",
-    "react-navigation-header-buttons": "^2.1.2",
-    "react-navigation-material-bottom-tabs": "^1.0.0",
     "react-redux": "^6.0.1",
     "recompose": "^0.30.0",
     "redux": "^4.0.1",
@@ -46,8 +53,8 @@
     "victory-native": "^31.0.0"
   },
   "devDependencies": {
-    "@babel/preset-flow": "^7.0.0",
     "@babel/core": "^7.5.0",
+    "@babel/preset-flow": "^7.0.0",
     "@babel/runtime": "^7.5.0",
     "@react-native-community/eslint-config": "^0.0.3",
     "babel-eslint": "^10.0.1",
