@@ -2,6 +2,7 @@ import { Provider } from 'react-redux';
 import React from 'react';
 import { View, ActivityIndicator, StyleSheet } from 'react-native';
 import { PersistGate } from 'redux-persist/integration/react';
+import { NavigationContainer } from '@react-navigation/native';
 import { colors } from './src/styles';
 
 import { store, persistor } from './src/redux/store';
@@ -11,17 +12,19 @@ import AppView from './src/modules/AppViewContainer';
 export default function App() {
   return (
     <Provider store={store}>
-      <PersistGate
-        loading={
-          // eslint-disable-next-line react/jsx-wrap-multilines
-          <View style={styles.container}>
-            <ActivityIndicator color={colors.red} />
-          </View>
-        }
-        persistor={persistor}
-      >
-        <AppView />
-      </PersistGate>
+      <NavigationContainer>
+        <PersistGate
+          loading={
+            // eslint-disable-next-line react/jsx-wrap-multilines
+            <View style={styles.container}>
+              <ActivityIndicator color={colors.red} />
+            </View>
+          }
+          persistor={persistor}
+        >
+          <AppView />
+        </PersistGate>
+      </NavigationContainer>
     </Provider>
   );
 }
