@@ -19,12 +19,12 @@ export default class GridsScreen extends React.Component {
       this.props.tabIndex
     ];
 
-  _openArticle = article => {
-    this.props.navigation.navigate({
-      routeName: 'Article',
-      params: { ...article },
-    });
-  };
+    _openArticle = article => {
+      this.props.navigation.navigate('Article', {
+        article,
+      });
+    };
+  
 
   renderRowOne = rowData => {
     const cellViews = rowData.item.map(item => (
