@@ -1,23 +1,37 @@
-/* eslint-disable import/no-unresolved */
-import React from 'react';
-import { Image, View, StyleSheet, Text } from 'react-native';
-import { createBottomTabNavigator } from 'react-navigation';
+import * as React from 'react';
+import { Text, View, Image, StyleSheet } from 'react-native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { colors } from '../../styles';
 
-import { colors, fonts } from '../../styles';
+import tabNavigationData from './tabNavigationData';
 
-import HomeScreen from '../home/HomeViewContainer';
-import CalendarScreen from '../calendar/CalendarViewContainer';
-import GridsScreen from '../grids/GridsViewContainer';
-import PagesScreen from '../pages/PagesViewContainer';
-import ComponentsScreen from '../components/ComponentsViewContainer';
+const Tab = createBottomTabNavigator();
 
-const iconHome = require('../../../assets/images/tabbar/home.png');
-const iconCalendar = require('../../../assets/images/tabbar/calendar.png');
-const iconGrids = require('../../../assets/images/tabbar/grids.png');
-const iconPages = require('../../../assets/images/tabbar/pages.png');
-const iconComponents = require('../../../assets/images/tabbar/components.png');
-
-const hederBackground = require('../../../assets/images/topBarBg.png');
+export default function BottomTabs() {
+  return (
+    <Tab.Navigator>
+      {tabNavigationData.map((item, idx) => (
+        <Tab.Screen 
+          key={`tab_item${idx+1}`}
+          name={item.name}
+          component={item.component}
+          options={{
+          tabBarIcon: ({ focused }) => (
+            <View style={styles.tabBarItemContainer}>
+              <Image
+                resizeMode="contain"
+                source={item.icon}
+                style={[styles.tabBarIcon, focused && styles.tabBarIconFocused]}
+              />
+            </View>
+          ),
+          tabBarLabel: ({ focused }) => <Text style={{ fontSize: 12, color: focused ? colors.primary : colors.gray }}>{item.name}</Text>
+        }} 
+        />        
+      ))}
+    </Tab.Navigator>
+  );
+};
 
 const styles = StyleSheet.create({
   tabBarItemContainer: {
@@ -35,130 +49,4 @@ const styles = StyleSheet.create({
   tabBarIconFocused: {
     tintColor: colors.primary,
   },
-  headerContainer: {
-    height: 70,
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    paddingBottom: 10,
-  },
-  headerImage: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    height: 70,
-    resizeMode: 'cover',
-  },
-  headerCaption: {
-    fontFamily: fonts.primaryRegular,
-    color: colors.white,
-    fontSize: 18,
-  },
 });
-
-export default createBottomTabNavigator(
-  {
-    Home: {
-      screen: HomeScreen,
-      navigationOptions: {
-        header: null,
-      },
-    },
-    Calendar: {
-      screen: CalendarScreen,
-      navigationOptions: {
-        header: (
-          <View style={styles.headerContainer}>
-            <Image style={styles.headerImage} source={hederBackground} />
-            <Text style={styles.headerCaption}>Calendar</Text>
-          </View>
-        ),
-      },
-    },
-    Grids: {
-      screen: GridsScreen,
-      navigationOptions: {
-        header: (
-          <View style={styles.headerContainer}>
-            <Image style={styles.headerImage} source={hederBackground} />
-            <Text style={styles.headerCaption}>Grids</Text>
-          </View>
-        ),
-      },
-    },
-    Pages: {
-      screen: PagesScreen,
-      navigationOptions: {
-        header: (
-          <View style={styles.headerContainer}>
-            <Image style={styles.headerImage} source={hederBackground} />
-            <Text style={styles.headerCaption}>Pages</Text>
-          </View>
-        ),
-      },
-    },
-    Components: {
-      screen: ComponentsScreen,
-      navigationOptions: {
-        header: (
-          <View style={styles.headerContainer}>
-            <Image style={styles.headerImage} source={hederBackground} />
-            <Text style={styles.headerCaption}>Components</Text>
-          </View>
-        ),
-      },
-    },
-  },
-  {
-    defaultNavigationOptions: ({ navigation }) => ({
-      // eslint-disable-next-line react/prop-types
-      tabBarIcon: ({ focused }) => {
-        const { routeName } = navigation.state;
-        let iconSource;
-        switch (routeName) {
-          case 'Home':
-            iconSource = iconHome;
-            break;
-          case 'Calendar':
-            iconSource = iconCalendar;
-            break;
-          case 'Grids':
-            iconSource = iconGrids;
-            break;
-          case 'Pages':
-            iconSource = iconPages;
-            break;
-          case 'Components':
-            iconSource = iconComponents;
-            break;
-          default:
-            iconSource = iconComponents;
-        }
-        return (
-          <View style={styles.tabBarItemContainer}>
-            <Image
-              resizeMode="contain"
-              source={iconSource}
-              style={[styles.tabBarIcon, focused && styles.tabBarIconFocused]}
-            />
-          </View>
-        );
-      },
-    }),
-    tabBarPosition: 'bottom',
-    animationEnabled: false,
-    swipeEnabled: false,
-    tabBarOptions: {
-      showLabel: true,
-      style: {
-        backgroundColor: colors.white,
-        borderTopWidth: 0.5,
-        borderTopColor: '#d6d6d6',
-      },
-      labelStyle: {
-        color: colors.grey,
-      },
-    },
-  },
-);
